Improved move script.

+++ xsd-extraction/move-to-src.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 
-rm -r ../src/main/java/eu
-rm -r ../src/main/java/uk
-rm -r ../src/main/java/de/vdv
+if [[ -d ./java ]]; then
+    rm -r ../src/main/java/eu
+    rm -r ../src/main/java/uk
+    rm -r ../src/main/java/de/vdv
 
-mv ./java/eu ../src/main/java/
-mv ./java/uk ../src/main/java/
-mv ./java/de/vdv ../src/main/java/de/
+    mv ./java/eu ../src/main/java/
+    mv ./java/uk ../src/main/java/
+    mv ./java/de/vdv ../src/main/java/de/
 
-rm -r ./java
+    rm -r ./java
+else
+    echo "Folder java does not exists. Aborting..."
+fi
